+++ app/src/main/java/dev/yasan/metro/tehran/domain/usecase/line/GetLineUseCase.kt
@@ -1,0 +1,12 @@
+package dev.yasan.metro.tehran.domain.usecase.line
+
+import dev.yasan.metro.tehran.domain.repository.line.LineRepository
+import javax.inject.Inject
+
+class GetLineUseCase @Inject constructor(
+    private val lineRepository: LineRepository
+) {
+
+    suspend operator fun invoke(lineId: Int) = lineRepository.getLine(lineId = lineId)
+
+}