+++ app/src/main/java/dev/yasan/metro/tehran/presentation/composable/screen/line/LineScreen.kt
@@ -1,0 +1,128 @@
+package dev.yasan.metro.tehran.presentation.composable.screen.line
+
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.requiredHeight
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.sharp.SwapVert
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.navigation.NavController
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import dev.yasan.kit.compose.foundation.grid
+import dev.yasan.kit.core.Resource
+import dev.yasan.metro.tehran.R
+import dev.yasan.metro.tehran.model.misc.Action
+import dev.yasan.metro.tehran.model.misc.LaunchSource
+import dev.yasan.metro.tehran.model.tehro.Line
+import dev.yasan.metro.tehran.model.tehro.Station
+import dev.yasan.metro.tehran.presentation.composable.common.teh.TehError
+import dev.yasan.metro.tehran.presentation.composable.common.teh.TehFooter
+import dev.yasan.metro.tehran.presentation.composable.common.teh.TehProgress
+import dev.yasan.metro.tehran.presentation.composable.common.teh.TehScreen
+import dev.yasan.metro.tehran.presentation.composable.screen.line.modules.StationItem
+import dev.yasan.metro.tehran.presentation.navigation.NavGraph
+import dev.yasan.metro.tehran.presentation.navigation.NavRoutes
+import dev.yasan.metro.tehran.presentation.theme.TehroIcons
+
+/**
+ * This screen shows the detailed data for a single [Line] which mainly includes the list of [Station]s inside it.
+ *
+ * @see NavGraph
+ * @see NavRoutes.routeLine
+ * @see NavRoutes.routeLineBase
+ */
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun LineScreen(
+    lineViewModel: LineViewModel,
+    navController: NavController,
+    lineId: Int,
+) {
+
+    val title = lineViewModel.title.observeAsState()
+    val lineColor = lineViewModel.lineColor.observeAsState()
+    val stations = lineViewModel.stations.observeAsState()
+
+    rememberSystemUiController().setStatusBarColor(color = lineColor.value ?: Color.DarkGray)
+
+    val orderAscending = rememberSaveable { mutableStateOf(true) }
+
+    LaunchedEffect(key1 = stations.value) {
+        if (stations.value is Resource.Initial) {
+            lineViewModel.loadStations(lineId = lineId)
+        }
+    }
+
+    val buttonAngle by animateFloatAsState(if (orderAscending.value) 180f else 0f)
+
+    TehScreen(
+        title = title.value ?: "",
+        color = lineColor.value ?: Color.DarkGray,
+        action = Action(
+            iconModifier = Modifier.rotate(buttonAngle),
+            icon = TehroIcons.SwapVert,
+            text = "",
+            onClick = {
+                orderAscending.value = !orderAscending.value
+            }
+        )
+    ) {
+
+        when (stations.value) {
+            is Resource.Error -> {
+                item {
+                    TehError {
+                        lineViewModel.loadStations(lineId = lineId)
+                    }
+                }
+            }
+            is Resource.Success -> {
+
+                val list = stations.value?.data ?: ArrayList()
+
+                val stationCount = list.size
+
+                item {
+                    Spacer(modifier = Modifier.requiredHeight(grid()))
+                }
+
+                items(
+                    items = list.apply {
+                        if (orderAscending.value) {
+                            sortBy { it.positionInLine }
+                        } else {
+                            sortByDescending { it.positionInLine }
+                        }
+                    },
+                    key = { it.id }
+                ) { station ->
+                    StationItem(
+                        modifier = Modifier.animateItemPlacement(),
+                        station = station,
+                        navController = navController,
+                        launchSource = LaunchSource.LINE
+                    )
+                }
+
+                item {
+                    TehFooter(text = stringResource(R.string.n_stations, stationCount))
+                }
+            }
+            else -> {
+                item {
+                    TehProgress()
+                }
+            }
+        }
+    }
+}