+++ app/src/main/java/dev/yasan/metro/tehran/presentation/preview/dbinfo/DatabaseInformationPreviewProvider.kt
@@ -1,0 +1,13 @@
+package dev.yasan.metro.tehran.presentation.preview.dbinfo
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import dev.yasan.metro.tehran.model.tehro.DatabaseInformation
+import dev.yasan.metro.tehran.util.PreviewHelper
+
+/**
+ * Provides preview parameters for [DatabaseInformation].
+ */
+class DatabaseInformationPreviewProvider : PreviewParameterProvider<DatabaseInformation?> {
+
+    override val values = sequenceOf(PreviewHelper.databaseInformation, null)
+}