+++ app/src/main/java/dev/yasan/metro/tehran/presentation/theme/Color.kt
@@ -1,0 +1,8 @@
+package dev.yasan.metro.tehran.presentation.theme
+
+import androidx.compose.ui.graphics.Color
+
+val themePrimary = Color(0xFF333333)
+val themePrimaryDark = Color(0xFF0c0c0c)
+val themePrimaryLight = Color(0xFF5c5c5c)
+val onThemePrimary = Color.White