+++ app/src/main/java/dev/yasan/metro/tehran/presentation/theme/Dimensions.kt
@@ -1,0 +1,5 @@
+package dev.yasan.metro.tehran.presentation.theme
+
+import androidx.compose.ui.unit.dp
+
+val dimenDivider = 2.dp