+++ app/src/main/java/dev/yasan/metro/tehran/ui/theme/Icons.kt
@@ -1,5 +1,0 @@
-package dev.yasan.metro.tehran.ui.theme
-
-import androidx.compose.material.icons.Icons
-
-val TehroIcons = Icons.Sharp